PP_BOT/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
jinja2==3.1.4
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.50.2
openpyxl==3.1.5
python-pptx==1.0.2
msal==1.31.1
